# Add benchmarks to check performance