Current values:
 - insert 100 docs with 0x0 queries: 271ms
 - insert 100 docs with 10x1 queries: 281ms
 - insert 100 docs with 1x10 queries: 256ms
 - remove 100 docs with 0x0 queries: 231ms
 - remove 100 docs with 10x1 queries: 272ms
 - remove 100 docs with 1x10 queries: 287ms
 - update 100 docs with 0x0 queries: 266ms
 - update 100 docs with 10x1 queries: 308ms
 - update 100 docs with 1x10 queries: 315ms